refactor command service
consolidate command.Datastore into command.Service
move redis implementation to a sub-package
